改好了删除item时的bugg

+++ AccountBook/app/src/main/java/com/zsf/accountbook/db/DatabaseHelper.java
@@ -64,9 +64,9 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         database.delete(COST,null,null);
     }
 
-    public void deleteOneData(long id){
+    public void deleteOneData(int id){
         SQLiteDatabase database = getWritableDatabase();
-        database.delete(COST,"id = ?", new String[]{Integer.toString((int) id)});
+        database.delete(COST,"id = ?", new String[]{Integer.toString(id)});
     }
 
     @Override
